requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.5
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=2.0
requests>=2.32.3
